requirements.txt
fastapi==0.115.0
uvicorn==0.40.0
pydantic==2.12.5
PyYAML==6.0.2
httpx==0.28.1
fastmcp==2.14.1
starlette==0.38.6
sse-starlette==1.6.1
psutil==5.9.8
